backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
supabase==2.0.0
python-dotenv==1.0.0
